% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{estimate}
\alias{estimate}
\title{Estimate Results Estimated for Multiple Dataset Replicates}
\usage{
estimate(data = NULL, formula, vars, estimator = estimatr::lm_robust, ...)
}
\arguments{
\item{data}{A list of dataset replicates produced with simdata().}

\item{formula}{A formula object specifying the right- and left-hand side variables to be used in estimation.}

\item{vars}{A vector or single character string indicating the names of covariates to return results for.}

\item{estimator}{The estimator to be used for estimation. Default is estimatr::lm_robust.}

\item{...}{Additional commands for the selected estimator function.}
}
\value{
The function returns a tidy dataframe of estimates for each selected variable for each dataset replicate.
}
\description{
This function allows users to iteratively estimate a linear regression for a list of dataset replicates produced with the simdata() function.
}
