% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{evaluate}
\alias{evaluate}
\title{Evaluate Power for Select Variables}
\usage{
evaluate(data, delta, level = 0.05)
}
\arguments{
\item{data}{The output from estimate().}

\item{delta}{A numeric vector of effect sizes over which to calculate power.}

\item{level}{The level of the statistical test. Default is 0.05.}
}
\value{
Returns computed power for each variable of interest over the range of use-specified effect sizes.
}
\description{
This function allows users to specify power to detect effects for select variables over a range of hypothetical effect sizes.
}
