% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seerrr-package.R
\docType{package}
\name{seerrr-package}
\alias{seerrr}
\alias{seerrr-package}
\title{seerrr: Simulate, Estimate, and Evaluate}
\description{
The seerrr package supports a simple and intuitive workflow for conducting computational power analysis.
}
\author{
\strong{Maintainer}: Miles Williams \email{milesdwilliams15@gmail.com} (\href{https://orcid.org/0000-0003-0192-5542}{ORCID})

}
\keyword{internal}
