% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate}
\alias{simulate}
\title{Iteratively Simulate a Data-Generating Process}
\usage{
simulate(R = 200, N = 500, ...)
}
\arguments{
\item{R}{The number of replicates to produce for a dataset. Default is 200.}

\item{N}{The sample size. Default is 500.}

\item{...}{User specified variables.}
}
\value{
The function returns a list of dataset replicates.
}
\description{
This function allows users to generate a dataset much like they would using a call to tibble(). However, rather than generate only a single dataset, the function returns a list of multiple replicates of a dataset.
}
